requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
apscheduler>=3.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
wasmtime>=23.0.0
